[7/10] Convert existing patterns to move_pop_back()
We already had quite a few places in our code where we use the
move-then-pop-back stuff we introduced in the previous patch. This patch
converts (=simplifies) those locations to use the new move_pop_back() function.

Actually in some of those locations we swapped the to-be-removed element with
the last element instead of only moving the last element over the to-be-removed
element. That's (a little bit) more work than needed. So also fixed now.