Restore mouse-joystick-emulation mode
When a MSX mouse is plugged in while holding the left mouse button pressed, it
will go into joystick emulation mode. This was implemented in openMSX, but
commit aae997be (2009/11/02) broke it. Thanks to FRS for reporting the problem.

Commit aae997be makes the input devices (joystick, mouse, trackball, ...) only
listen to input events when they are plugged in. Though what I forgot was that
immediately when plugging in the mouse it already needs to know the status of
the left mouse button. Fixed by explicitly querying the status on plug (instead
of relying on events).