Fix control-logic in BeerIDE
BeerIDE uses a PPI (I8251) to control an IDE device. Port A and B of the PPI
are used to read/write the (16-bit) IDE databus. Port C is used to set some IDE
control lines (/IORD, /IOWR, /CS0 and A2-A0). PPI ports A and B are dynamically
reconfigured as input- or as output-ports. PPI port C is always configured as
output, both the high and the low part (2x4-bits) of port C.

Because in principle the direction (input/output) of the high and the low part
of port C can be set differently, in openMSX there are two separate methods
(writeC0() and writeC1()) to notify a device for changes on the C port. So when
both parts are configured as output (as is the case for BeerIDE) and the MSX
writes a new 8-bit value, both these writeCx() methods are called. It happens
that first writeC1() is called followed by writeC0().

In case of BeerIDE, the control signals /IORD and /IOWR are written via
writeC1() and the address signals A2-A0 are written via writeC0(). Now when
both the control signals and the address signals change at the same time the
following happens (scenario before this patch):
 - First writeC1() is called, we check /IORD and /IOWR and we possible execute
   a read or write on the IDE device. This uses the _old_ address.
 - Shortly after writeC0() is called. This latches the _new_ address, but does
   _not_ re-execute the IDE read or write.

The IDE standard probably has timings-diagrams for when the A2-A0 signals are
allowed to change wrt the /IORD and /IOWR signals. And most likely the
situation we encounter above is not permitted (I didn't explicitly check the
IDE standard though). The code in the Beer-ROM is usually very careful to avoid
such timing glitches: it very carefully only changes either the upper or the
lower part of port C, and only changes the address signals when no read or
write is active. Though in the initialization code there is a bug(?) that does
change both at once.

I fixed this by also (re-)executing the IDE read/write when the address lines
change. This is likely also how real HW _could_ react to such glitches, and it
does seem to fix the geometry-detection problem during boot (using Beer Rom
version 1.8).

I changed some member variables, but I didn't bother keeping the savestates
backwards compatible: the device is very new, and never properly worked yet. So
it's unlikely there already are useful savestates for it.