Fixed OSD-widget name management bug
This regression was introduced in revision 7b45c86c. That patch replaced the
name-lookup hash_maps located in each OSDWidget with only one hash_map located
in OSDTopWidget.

The problem occurred when destroying a widget that still had children. Those
children are also destroyed, but we forgot to remove their names from the top
map. Before 7b45c86c this wasn't a problem because the sub-maps that contains
the names of the children itself disappear.

Fix is easy enough: when removing a name, also (recursively) remove the names
of the children.

Though to be able to implement this efficiently (O(number-of-children)) I had
to give back each OSDWidget a name member (revision 7b45c86c removed that
member). This may seem like a step-back, but actually it doesn't make any
difference: we add the name to the OSDWidget, but remove it from the hash_map
in OSDTopWidget (so turn hash_map into hash_set).