Change set->vector return type in parseRange()
The parseRange() function is e.g. used to parse the numeric stuff in:
  record_channels SCC 1,3-5
In this case the range of allowed values is quite limited (1-5). And this is
true in all cases where we use parseRange().

So it's simpler/faster to use a (sorted) vector instead of a set to return the
result. See previous commit for a more detailed motivation.