Fixed 'Hardware... > Connectors...' OSD sub-menu
Went wrong when there were connector names that contain spaces. For example
"Generic MSX-Audio-MIDI-in".

The problem was that we tried to split the string
   Generic MSX-Audio-MIDI-in: --empty--
with the Tcl command 'split [plug $connector] ": "'. Though instead of
splitting on the sequence ": " it splits on either ' ' or ':'. So this would go
wrong on any connector/pluggable names that contain ' ' or ':' characters.
Solved by explicitly searching on the sequence ": " and taking the substring
past that point.

I quickly reviewed all other uses of the split command. They all seem fine.