[1/3] return 'MemBuffer<uint8_t>' instead of string
This patch series fixes a few inefficiencies in hex or base-64
encoding/decoding. This is used for storing binary blobs (e.g. memory mapper
content) in XML savestates or for storing initial memory patterns in machine
configuration files.

--

In this first step we change the decode() functions in the HexDump and Base64 classes
to return a MemBuffer<uint8_t> instead of a std::string. The latter was
inefficient because:
 - (in Base64) we first needlessly zero-initialize the memory
 - (in HexDump) we reallocate the memory a few times (while growing)
Both could be fixed by using std::string::reserve(), but using
MemBuffer<uint8_t> instead of std::string is an equally good solution and it
more closely matches what's actually meant (return a buffer of bytes).

To be more consistent, the input side (the encode() function) now takes a
uint8_t* as input instead of a void*.