[1/5] Extend OutputBuffer with insert_tuple_ptr()
First some background info:

Looking back, I'm quite happy with how the serialization code in openMSX turned
out. The user has to write one simple method like:

    template<typename Archive>
    void MyClass::serialize(Archive& ar, unsigned version) {
        ar.serialize("foo", foo);
        ar.serialize("bar", bar);
    }

And from this single templatized method 4 instances are generated: 2 to
save/load to/from an XML file and 2 to save/load to/from a memory buffer. The
in-memory variants are _much_ faster than the XML variants (e.g. because they
don't need to bother with tags or with backwards compatibility) and are used to
take regular snapshots for the reverse system.

Now concentrating on the in-memory variants. When looking at the generated
code, I'm very happy with the load-from memory version, that code is close to
optimal (e.g. to restore an integer value, it loads 4 bytes from the buffer,
stores those into the member and increments the pointer. So only 3 x86
instructions, hard to do much better).

The generated code to save-to-memory is also reasonably good, but there's one
thing that has been bothering me for a long time. Let's explain with an
example. For the serialize() function above the generated code would do more or
less the following:
  - ar.serialize("foo", foo)
     -> make sure there's room for 'sizeof(foo)' extra bytes in the output,
        so do some pointer comparisons and possible grow the buffer
     -> copy 'foo' to the output buffer
     -> adjust the output pointer
  - ar.serialize("foo", bar)
     -> and exactly the same as above for the 'bar' member

A more efficient implementation would do:
   -> make sure there's room for 'sizeof(foo) + sizeof(bar)' extra bytes
   -> copy both 'foo' and 'bar' to the output
        (if the 'foo' and 'bar' members happen to be adjacent in memory (and
        suitably aligned) a smart compiler could even coalesce both load/store
        pairs into a single pair)
   -> adjust the output pointer once

Checking whether there's enough room is not very expensive in absolute terms.
But relative to doing the actual data move it is. By doing the above
optimization, saving to memory can become roughly twice as fast.

In the past I already made several (failed) attempts to teach the
serialization code to perform the above optimization. In this patch series I
finally, partly, succeed. Though I'm using variadic templates, a c++11 feature
that wasn't available yet at the time I originally wrote this code.

--

This patch only does some preparatory work. It adds a method insert_tuple_ptr()
to the OutputBuffer class (this class represents the buffer where in-memory
save states write to). This method allows to put a bunch of elements into the
buffer using a single call, and this way the OutputBuffer only needs to
check/grow once for available space. This new method is not yet used in this
patch.

The patch also renames some 'byte' types to the standard 'uint8_t' type and
cleans up some #includes.