[2/5] Added variadic serialize(...) functions
Let's revisit the example I gave in the previous commit message

    template<typename Archive>
    void MyClass::serialize(Archive& ar, unsigned version) {
        ar.serialize("foo", foo);
        ar.serialize("bar", bar);
    }

The optimization we want should be done over the two serialize() calls. That's
hard (my past attempts to do this failed). But if we use c++11 variadic
templates we can rewrite the above code as

    template<typename Archive>
    void MyClass::serialize(Archive& ar, unsigned version) {
        ar.serialize("foo", foo,
                     "bar", bar);
    }

More general, this new version of serialize() accepts not just one tag-variable
pair, but an undefinite number of pairs. And now the implementation of
serialize() has a full view of what should be done.

In case of the save/load to/from XML, the variadic serialize() function does
_exactly_ the same as repeated calls with single pairs. So for those nothing
changes. For saving to memory we do the following:
 - We walk over the list of pairs (walk over all arguments in steps of two).
 - If the current pair can be 'optimized' (means serialized with a memcpy) we
   don't process it just yet, but only remember it in a tuple of
   still-to-be-processed elements.
 - If the current pair cannot be optimized we handle it immediately in exactly
   the same way as before.
 - When we've walked over all parameters we process to tuple of collected
   elements with a single call to insert_tuple_ptr() (see previous patch).

Some remarks:
 - The order in which the elements are stored has changed compared to before
   this patch. So loading from memory has to be changed accordingly.
 - While walking over the elements we build a tuple of still to be processed
   elements. This stuff is _completely_ optimized away by the compiler. So
   after optimization it's as-if the element pairs were immediately present in
   a suitable order.