[4/5] Improve serialization of array's
After the changes in this series so far I looked at the quality of the
generated code. There were still a few things that could be tweaked so that the
reserve-memory-only-once optimization can trigger more often. That's done in
this and in the following patch.

Before this patch serializing an array (e.g. an array of 4 int's) was done in a
small loop. Though it's more efficient to just do one memcpy(). This also
enables the serialization of the array to be merged with the serialization of
other simple elements.