[5/5] Tweak serialization of Clock EmuTime EmuDuration FixedPoint
The data types Clock, EmuTime, EmuDuration and FixedPoint internally contain
only a single (32 or 64 bit) integer. Serializing them to memory can be done
with a simple memcpy(). This patch makes the serialization framework aware of
this. And that allows to optimize serialization of these types better (e.g.
group them together with other simple data types).

--

At the end of this patch series, the generated code looks a lot better. The
overhead of the reverse system has measurably decreased, but only by a small
amount (~1%). The reason is that by far most time of creating an in-memory
savestate is spend in compressing the (V)RAM, not in storing (several hundred)
individual small fields. I also have ideas to improve that part (storing the
RAMs), but I need more time to actually implement it.