Investigate CALL instruction on R800
I did some measurements on R800 for the CALL (and related) instructions. This
was triggered by Grauw's findings some time ago (see
http://map.grauw.nl/resources/z80instr.php for his results).

This patch only adds the description/results of the measurements. It doesn't
make any changes in the emulation yet (so that openMSX matches these results).
That's something for after the 0.12 release.