Fix(?) LessTupleElement for vs2015
Visual studio 2015 triggered errors like
    C2338: tuple_element index out of bounds
on compiling our LessTupleElement utility. Hopefully this patch fixes that.

--

It's not 100% clear whether this was a bug in the vs2015 library or whether our
code relied on a non-standardized implementation detail. Here's some background
info:

Part of the old implementation of LessTupleElement went like this:

    template<int N> struct LessTupleElement {
        template<typename TUPLE>
        bool operator()(const TUPLE& x, const TUPLE& y) const { ... }

        template<typename TUPLE>
        bool operator()(const typename std::tuple_element<N, TUPLE>::type& x,
                        const TUPLE& y) const { ... }
        template<typename TUPLE>
        bool operator()(const TUPLE& x,
                        typename std::tuple_element<N, TUPLE>::type& y) const { ... }
    };

So we have a LessTupleElement struct with three overloaded function-call
operators. The 1st takes 2 tuples, the latter 2 take a tuple and a parameter
with as type the N-th element of that tuple. We rely on SFINAE to deduce the
correct type for the 'TUPLE' template parameter and select the correct
overload. E.g. when you call this functor with a 'int' and a 'tuple<int,
string>' parameter the 2nd overload should be selected and TUPLE should be
deduced to be 'tuple<int, string>'. Also in this example it's impossible to
find a type for TUPLE so that the 1st and 3rd overload don't generate some
error (and we actually do want these to be SFINAE failures).

At least this is the story for gcc (libstdc++), clang (libc++) and visual
studio up-to vs2013. In visual studio 2015 the tuple_element utility is
implemented so that it can be instantiated with any type. For 'correct'
instantiations it gives the expected result. For all other instantiations it
triggers a static_assert 'tuple_element index out of bound'. Often this is a
useful error message (e.g. when asking for the 4th element in a tuple with only
two elements), but in our case it prevents SFINAE from working (because
tuple_element can now be instantiated with anything).

So again, it's not clear to me whether the visual studio 2015 implementation is
conforming to the c++ standard. According to this it isn't
    http://en.cppreference.com/w/cpp/utility/tuple/tuple_element
but that's not an authoritative source. And I didn't check the actual standard.

--

This patch changes the implementation of our LessTupleElement utility to no
longer require std::tuple_element. Instead we now need c++11 variadic
templates. Fun fact: LessTupleElement was originally implemented without
variadic templates because an older version of visual studio didn't support
those yet.