[3/10] Use rfind_(if_)unguarded()
Suppose we have a vector, we search an element using linear search (e.g.
because vector is unordered or very small) and we then erase that element.
Typical code to do this is:
	auto it = find(begin(v), end(v), e);
	if (it != end(v)) v.erase(it);
Or if we are sure the to-be-removed element is certainly present we can use
this faster variant:
	v.erase(find_unguarded(v, e));

Now think about what's happening. We start at the front of the vector and walk
over all elements until we found the element we're interested in. Then we erase
that element, that means we shift all later elements one position to the front.
So this means that wherever the element was located (near the front, middle or
back) we always touch the memory locations for the whole vector.

Now let's make a small change to the example (literally only add 1 character):
	v.erase(rfind_unguarded(v, e));

This also does a linear search, but it starts from the back instead of the
front. When we arrive at the element we again erase it by moving all later
elements one position to the front. But now we move elements that we've just
visited during the search, so they are almost guaranteed to be in the CPU's
L1-cache.

If we assume the elements are unordered and/or we search for a random element
there's no reason to prefer searching from front-to-back or from back-to-front.
However a pattern that often occurs in practice is that elements are
added/removed in a (more or less) FIFO order. So the to-be-removed element is
more likely to be located near the back of the vector. In that case starting
the search from the back has an additional advantage.

--

So this patch replaces all erase(find_unguarded(..)) patterns in our code with
erase(rfind_unguarded(..)).