[9/10] Tweak removing 'garbage boards'
For some reason, in the Reactor, code we postpone deleting of MSXMotherBoard's
(see 1aa4ba50 for details). The do this we put the to-be-deleted board in the
garbageBoards vector and send a OPENMSX_DELETE_BOARD event. When that event
(eventually) gets handled we actually delete the board.

It is possible that multiple boards are scheduled to be deleted before the
DELETE_BOARD event handler gets executed. Before we made sure that each event
deletes the board for which it was scheduled. Though the exact order in which
the boards get deleted doesn't matter (as long as they all get deleted).

This patch takes advantage of this freedom and slightly optimizes this deletion
step: it's cheaper to remove the last element from a vector than the first
element.