[6/10] Added move_pop_back() function
This utility removes an element from a vector by
 - first moving the last element of the vector over the to-be-removed elememt
 - then erase the last (now unused) position
In case the order of the elements in the vector doesn't matter, this is a
faster alternative compared to calling the vector::erase() method (because that
one has to shift all later elements to the front).

This patch only adds the new function but doesn't use it yet.