[1/10] Document sorting criteria (if any) for vectors
This might be the most important patch in this series even though it doesn't
make any actual code changes. It only adds some comments.

This patch documents for many of the std::vector variables we use whether the
elements they contain are kept in a particular order. And if so what the
sorting criteria is.

Later patches in this series tweak some of the operations we perform on vectors,
but some of these tweaks are only allowed if e.g. the order of the elements
doesn't matter.