[4/10] Added partition_copy_remove() utility function
This takes a range of elements and a predicate as input. It moves all elements
of the range that satisfy the predicate to a given output range and all
elements that don't satisfy the predicate are moved in-place to the front of
the original input range. See comments in code for more details.

This patch only adds the new function but doesn't use it yet.