[3/3] decode inplace
The first patch in this series changed the return type of the decode()
functions from std::string to Membuffer<uint8_t>. In both cases the decode()
function allocates a new memory block for the decoded data. In some (but not
all) cases the only thing we do with this memory block is copy it to another
buffer and deallocate the returned memory block. It's more efficient to
directly decode the data to the final destination.

So this patch adds an extra decode_inplace() function (the existing decode()
function is kept as well). This new function takes a pointer to the destination
buffer and the size of this buffer. It returns an error if the size of the
decoded data is not exactly the same as the size of the destination buffer.
This error check has a cost (because we only know the size of the decoded block
after we've decoded it, and during decoding we cannot accidentally write past
the end of the destination buffer), but this cost is less than the cost of
allocate+copy+deallocate.

So far this new decode_inplace() function is  only used when loading blobs from
XML savestates.