[1/3] Enable/fix -Wshadow compiler warnings
This patch enables -Wshadow and then fixes all the resulting compiler warnings.
Both for gcc and clang.

The motivation for this patch is the following: when compiling with visual
studio 2015 we currently get about 1000 compiler warnings. We want to get this
list back down to a more reasonable number. A large portion of these new
warnings are "C4458: declaration of 'parameter' hides class member". These more
or less correspond to gcc's -Wshadow warnings.

Personally I'm not sure yet whether I like this warning or not. I can see how
it can point to confusing or even buggy code (though I didn't spot any bugs
when creating this patch). But it also forces renames in otherwise perfectly
clear code. I suggest we try it out for a few months, if it's not useful we can
always disable it again.