[2/3] Fix(?) vs2015 C4307 compiler warning
Visual studio 2015 generated _a_lot_ of duplicated warnings for
   c:\compiler\openmsx\src\utils\xxhash.hh(88): warning C4307: '+': integral constant overflow
And the compiler is correct, there is an overflow in that line. But it is also
how the code is intended (we want 32-bit modulo arithmetic, and that's what the
c++ standard guarantees for unsigned types).

It's also relatively easy to work around this warning. And after the compiler's
optimizer has done it's job, the generated code should be identical to the
generated code before this patch.