[2/2] Tweak string comparisons in RomDatabase parsing
During rom database parsing we sometimes already know the first character of a
tag and still need to check the remaining characters. E.g. we know that the tag
starts with a 't' and still need to check whether the full name might be
"type". We now have a choice, we can either compare the full tag name with
"type" or compare the tag minus the first character with "ype".

In the last patch we introduced the small_compare() function. This functions
runs a bit faster if the to-be-compared string has length 1, 2, 4 or 8. And for
strings of length 5, 6 or 7 it runs equally fast (so 5 is not faster than 7).

This patch combines the ideas of the previous two paragraphs. So case-by-case
it decides it check the full tag or the tag minus the first character. I
verified that the generated code is indeed simpler (~100 bytes, ~20
instructions shorter). But I didn't re-measure the startup time.