Migrate Tcl to 8.5.18 in 3rdparty system.
This vastly reduces the needed patches (contacting the Tcl developers with the
patches we are applying helped a great deal, as they all landed in their code
now). The only one needed is for cross compiling. This patch was explicitly
refused by the Tcl developers (well, it's a hack anyway).

This version should also compile fine on Visual Studio 2015.