Test: remove manually created move constructor/assignment
According to the C++11 standard, the compiler should automatically create the
move constructor and move assignment operator for this class. And it does
exactly that with gcc and clang. However vs2013 didn't do this correctly and as
a workaround we had to manually define these operations. Now we've switched to
vs2015 and this problem *should* be fixed.

As a test I've removed the workaround in only one class. Let's see what the
vs-build-bot reports. If it's OK we can remove similar workarounds from various
other places.