Revert some problematic patches for vs2015
This reverts commits
   633507c2922f222ff6c4aeeca4e9ccc334a06302
   0a00e8a1256e7ce3ea85ee611c3762d3de159ca6
   0050c9688f2d73a338bfb2849fdedc226b9aa669
   ddf6bda1112bb662f8293baab3a8ee4a1adffe86
   345abade7a7231b2ee249e62b541adb18989b9b8
   25bf9193025aacef763e849b72e32aae1413b70f

The above patch series made some improvements to the serialization framework so
that creating an in-memory savestate became a bit faster. For more details see
those commits.

However these changes are more demanding on the compiler and vs2015 seems to
have a hard time compiling/linking them. We already had to switch to the 64-bit
vs2015 tools (a good idea anyway) to work around a 4GB address space limit. But
even then sometimes (not always?) they bring down Vampier's compile-server.

The above patch series brings an improvement of about 1% on the total emulation
time (when reverse is enabled). That's nice, but maybe not super important. So
until the vs2015 crashes are resolved I'm reverting the patch series.

I did create a new git branch 'vs2015-serialize' where I did not revert these
patches. This way we can easily verify test new vs releases and re-apply the
patches to the master branch when the problem is solved.