Fix parsing of '&#x12AB;' in AdhocCliCommParser
See here for a scenario in which the bug manifested itself:
  http://www.msx.org/forum/msx-talk/openmsx/catapult-doesnt-always-change-disk-image-linux

The bug was that I mistakenly thought &#123; should be interpreted as
hexadecimal instead of decimal (and I didn't know about &#x123;). The unittest
for the parser contains the same mistake, so this confirms it's a design-
rather than an implementation-error.

Once understood, fixing the bug was easy.