Feature request #984: Printer port debuggable
Implement feature request #984: add a debuggable for the printer port data and
strobe.

There's a debuggable with the same name as the printerport (in theory it's
possible to emulate more than 1 printerport). The debuggable has two addresses:
 0: strobe (only bit 0 has meaning)
 1: data (8 bit)
This corresponds to the location of this information when writing to IO port
0x90-0x91.