Improve CliServer's destructor
Do not join the accept thread if it was never started or if waking it
failed. While it is risky to continue destruction with the thread
still having a reference to the CliServer object, waiting for the
thread to exit would hang openMSX, which is certainly bad.

Do not call Thread::stop(), since that method will be going away. Also
there is no telling what it will do if the thread is inside accept().