Improve exitLoop handling
Mark exitLoop as 'volatile' to make it explicit that we're sharing
it between threads and to prevent the compiler from reordering the
statements that accesses it.

Check exitLoop's value immediately after accept(), to decrease the
chance of crashing if clean shutdown of the accept thread by the
destructor failed.

Made exitLoop static, so it is not an error when accessing it after
the CliServer object was destructed. Note that sock_startup() and
sock_cleanup() are not refcounted, so we can only deal with one
CliServer instance at any time anyway.

Note that there is still a race condition when CliServer is destructed
just after the accept thread checks exitLoop.