Return array_ref<uint8_t> from File::mmap()
Before we returned a pointer to the buffer via the return value and the size of
the buffer via an output-parameter. Now we return both via a single return
value.

This raises the abstraction level without sacrificing performance. And it
enables some (small) cleanups. It was inspired by this talk:
  https://isocpp.org/blog/2015/09/bjarne-stroustrup-announces-cpp-core-guidelines

(Another cleanup patch I had lying around for a few years but for some reason
forgot to commit.)