Don't call 'deliverEvents' while reversing.
This was done to have the reverse bar being updated nicely while doing a long
time warp in reverse. But on second thoughts: this is really not safe. It also
means that things like hotkeys which can trigger new commands are active whilst
doing reverse. It could mean a new goTo command while performing a goTo (which
is what I often saw, causing me to investigate the error message that was the
result of that), but even worse: it could even mean the machine will get
deleted, so that reverse will be working with a deleted object. This will
obviously crash (and I confirmed that in a test).

In other words: our code is not prepared to handle the consequences of
'deliverEvents' during a reverse action. So better not do that, even though it
means the UI experience of the reverse bar will deteriorate with this again. At
least the progress will still function, which is already a huge improvement.