TTH got recalculated (once) on reverse
Scenario to reproduce the problem:
- Execute: dd if=/dev/zero of=/tmp/testbin bs=1024 count=1MB
    The step isn't strictly needed. It creates a big HD image so that
    calculating a full hash takes enough time so that openmsx shows progress
    information. The problem also occurs for small HD images, but then it's
    less noticable.
- Execute: openmsx -ext ide -hda /tmp/testbin
- OpenMSX boots and calculates the initial hash for the HD.
- Next execute a reverse action (e.g. press 'PgUp').
- OpenMSX re-calculates the hash. Since this calculation is quite heavy
  (multiple seconds) we should try to avoid it.
- However reversing again does _not_ retrigger the calculation.

This bug was caused by a wrong cache-validity-check. Checking the file
modification timestamp also updates this timestamp (to avoid having to query it
twice). But the initial validity check skipped the timestamp check (for no good
reason), so the timestamp was only correctly initialized on the second check.

Thanks to Manuel for discovering and analyzing this problem. I merely took his
suggested solution and tweaked it a little.