Fixed regression introduced by the VDP delay refactoring
Commit 8b14c7ef1 moved handling of the VDP IO delay from MSXMatsushita (and
part of MSXS1990 in a follow-up commit) to the VDP class. This refactoring
could trigger some asserts. Maybe most easy to reproduce by booting a turbor
with a non-updated config file.

The problem was located in the detailed EmuTime accounting. After calling
MSXCPU::waitCycles() the time has moved forward, so we must make sure to also
use this new time point for the following calculations (e.g. in
VDP::writeIO()). Also when waiting for a few cycles there's a chance there's a
syncpoint located in that (small) time interval. So we must make sure to
execute those sync points.

To fix this problem, and also to make sure we don't make similar mistakes in
the future, I've changed the prototype of the MSXCPU::waitCycles() method. It
now takes an EmuTime as input parameter (the current time) and it returns an
EmuTime (the updated time point).