Fix border drawing in text mode with YJK/YAE bits
This fixes the impossible screen 12 remnant described in this bug report:
  https://github.com/openMSX/openMSX/issues/1006

The problem was that the DisplayMode::isTextMode() method returned false for
text modes that had the YJK or YAE mode bit set. Because of this the border
sizes were calculated as 'narrow' (text mode has wider horizontal border
compared to character or bitmap display modes), but the actual display pixels
were drawn using the text-drawing routines. So some pixels were not drawn at
all and the previously visible data was still shown (part of a screen 12 image
in the bug report).

This fixes the remnant pixels, but the image does not yet look the same as on a
real MSX. OpenMSX does not yet emulate the (bogus) text+YJK modes. Actually
none of the non-screen8+YJK bogus modes is emulated.