Fix #1014: Baltak Rampage: characters in greetings part are one pixel offset
The V9938 datasheet states that for a LMMC command the first to-be transfered
value (from CPU->VRAM) must be written to VDP register R#44 before the command
starts. This demo does not do this, instead it just starts the command and only
then writes all values to R#44. It also starts a new LMMC command (writes
R#36,38,40,42,46) before the previous command is finished. I'm not 100% sure,
but I think it's also important to mention that no other commands are issued
in between all LMMC commands (or at least there are no other writes to R#44).

Before this patch, when openMSX starts executing a LMMC command, it immediately
uses the last written value to R#44 as a valid input. It does this by forcing
the internal 'transfer' boolean to true. Normally this boolean is set to true
on a write to R#44 and reset to false when the LMMC (or HMMC) command consumes
the value.

This patch removes the 'force to true' part. When you follow the recommended
LMMC recipe that's described in the V9938 datasheet, the behavior stays the
same because the initial R#44 write does also sets transfer=true. Alex Wulms'
original command engine (still used in blueMSX) also implements it like this.

So 'Baltak Rampage' doesn't follow this recipe, and because of this I think the
very first LMMC command will go wrong (also on a real machine). Though this is
only visible for a very short time (if at all). And it does make using the LMMC
a lot simpler (on V9990 the equivalent command is changed to first start the
command and only then transfer the first value .. much simpler to program).