Improve title guessing for unknown ROMs.
Fixes #1012.

This also fixes the assumption of rom_info that the title guesser returns a
device name. This has now been made explicit. Internally, guess_title now also
first guesses ROM device names and later checks whether this is just a file
name. If it is, it beautifies it by stripping of known file extensions and
making it a title.