Fix memory bug in MSXMotherBoard::userNames
To trigger it, type in the openMSX console:
1) carta <your-favorite.rom>
2) carta <your-favorite.rom>     (same rom again)
3) exit
Sometimes this triggers a crash/assert (depending on the build flavour). It's
easier to trigger when there's more time between step 1) and 2).

The problem was that the 'userNames' map used a 'string_ref' instead of a
'string' as key. In practice this bug is usually harmless because very often
the string really has a lifetime longer than the lifetime of the map. Not so in
the above scenario.